+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 Notable changes to the project will be added here.
 
+# UNRELEASED
+
+## Added
+- Progressbar for both syncing live stacks and syncing stack picks.
+
 
 # 1.0.0 - 2020-08-01